support zio-RC18-2
