+++ build.sbt
@@ -27,21 +27,23 @@ inThisBuild(
 addCommandAlias("fmt", "all scalafmtSbt scalafmt test:scalafmt")
 addCommandAlias("check", "all scalafmtSbtCheck scalafmtCheck test:scalafmtCheck")
 
+val zioVersion = "1.0.0-RC18-2"
+
 lazy val `zio-ftp` = project
   .in(file("."))
   .settings(stdSettings("zio-ftp"))
   .settings(
     libraryDependencies ++= Seq(
-      "dev.zio"                  %% "zio"                     % "1.0.0-RC17",
-      "dev.zio"                  %% "zio-streams"             % "1.0.0-RC17",
+      "dev.zio"                  %% "zio"                     % zioVersion,
+      "dev.zio"                  %% "zio-streams"             % zioVersion,
       "com.hierynomus"           % "sshj"                     % "0.27.0",
       "commons-net"              % "commons-net"              % "3.6",
       "org.scala-lang.modules"   %% "scala-collection-compat" % "2.1.4",
       "org.apache.logging.log4j" % "log4j-api"                % "2.12.1" % Test,
       "org.apache.logging.log4j" % "log4j-core"               % "2.12.1" % Test,
       "org.apache.logging.log4j" % "log4j-slf4j-impl"         % "2.12.1" % Test,
-      "dev.zio"                  %% "zio-test"                % "1.0.0-RC17" % Test,
-      "dev.zio"                  %% "zio-test-sbt"            % "1.0.0-RC17" % Test
+      "dev.zio"                  %% "zio-test"                % zioVersion % Test,
+      "dev.zio"                  %% "zio-test-sbt"            % zioVersion % Test
     ),
     testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework")
   )
@@ -54,7 +56,7 @@ lazy val docs = project
     scalacOptions -= "-Yno-imports",
     scalacOptions -= "-Xfatal-warnings",
     libraryDependencies ++= Seq(
-      "dev.zio" %% "zio" % "1.0.0-RC17"
+      "dev.zio" %% "zio" % zioVersion
     ),
     unidocProjectFilter in (ScalaUnidoc, unidoc) := inProjects(`zio-ftp`),
     target in (ScalaUnidoc, unidoc) := (baseDirectory in LocalRootProject).value / "website" / "static" / "api",
